cleanup

+++ lib/detox/main.py
@@ -208,7 +208,8 @@ class Detox(object):
             self.history.close_deletion_run(run_number)
 
         finally:
-            os.remove(detox_config.activity_indicator)
+            if os.path.exists(detox_config.activity_indicator):
+                os.remove(detox_config.activity_indicator)
 
         logger.info('Detox run finished at %s\n', time.strftime('%Y-%m-%d %H:%M:%S'))
 
@@ -292,7 +293,8 @@ if __name__ == '__main__':
     from common.transaction import TransactionManager
     from common.demand import DemandManager
     import common.interface.classes as classes
-    from detox.rules import ActionList, BelongsTo
+    from detox.policies.policies import ActionList
+    from detox.policies.site import BelongsTo
 
     parser = ArgumentParser(description = 'Use detox')
